Introducing test GA algorithm

+++ src/main/java/models/LevelData.java
@@ -2,12 +2,12 @@ package models;
 
 public class LevelData {
 
-    private int hp;
-    private int heal_5s;
-    private int attack;
-    private int defense;
-    private int sp_attack;
-    private int sp_defense;
+    private double hp;
+    private double heal_5s;
+    private double attack;
+    private double defense;
+    private double sp_attack;
+    private double sp_defense;
     private String move_speed;
     private double cdr;
     private double life_steal;
@@ -15,12 +15,12 @@ public class LevelData {
     private double tenacity;
     private double attack_speed;
 
-    public LevelData(int hp,
-                     int heal_5s,
-                     int attack,
-                     int defense,
-                     int sp_attack,
-                     int sp_defense,
+    public LevelData(double hp,
+                     double heal_5s,
+                     double attack,
+                     double defense,
+                     double sp_attack,
+                     double sp_defense,
                      String move_speed,
                      double cdr,
                      double life_steal,
@@ -42,51 +42,51 @@ public class LevelData {
         this.attack_speed = attack_speed;
     }
 
-    public int getHp() {
+    public double getHp() {
         return hp;
     }
 
-    public void setHp(int hp) {
+    public void setHp(double hp) {
         this.hp = hp;
     }
 
-    public int getHeal_5s() {
+    public double getHeal_5s() {
         return heal_5s;
     }
 
-    public void setHeal_5s(int heal_5s) {
+    public void setHeal_5s(double heal_5s) {
         this.heal_5s = heal_5s;
     }
 
-    public int getAttack() {
+    public double getAttack() {
         return attack;
     }
 
-    public void setAttack(int attack) {
+    public void setAttack(double attack) {
         this.attack = attack;
     }
 
-    public int getDefense() {
+    public double getDefense() {
         return defense;
     }
 
-    public void setDefense(int defense) {
+    public void setDefense(double defense) {
         this.defense = defense;
     }
 
-    public int getSp_attack() {
+    public double getSp_attack() {
         return sp_attack;
     }
 
-    public void setSp_attack(int sp_attack) {
+    public void setSp_attack(double sp_attack) {
         this.sp_attack = sp_attack;
     }
 
-    public int getSp_defense() {
+    public double getSp_defense() {
         return sp_defense;
     }
 
-    public void setSp_defense(int sp_defense) {
+    public void setSp_defense(double sp_defense) {
         this.sp_defense = sp_defense;
     }
 
@@ -102,7 +102,7 @@ public class LevelData {
         return cdr;
     }
 
-    public void setCdr(int cdr) {
+    public void setCdr(double cdr) {
         this.cdr = cdr;
     }
 
@@ -110,7 +110,7 @@ public class LevelData {
         return life_steal;
     }
 
-    public void setLife_steal(int life_steal) {
+    public void setLife_steal(double life_steal) {
         this.life_steal = life_steal;
     }
 
@@ -118,7 +118,7 @@ public class LevelData {
         return crit_chance;
     }
 
-    public void setCrit_chance(int crit_chance) {
+    public void setCrit_chance(double crit_chance) {
         this.crit_chance = crit_chance;
     }
 
@@ -126,7 +126,7 @@ public class LevelData {
         return tenacity;
     }
 
-    public void setTenacity(int tenacity) {
+    public void setTenacity(double tenacity) {
         this.tenacity = tenacity;
     }
 
@@ -134,7 +134,7 @@ public class LevelData {
         return attack_speed;
     }
 
-    public void setAttack_speed(int attack_speed) {
+    public void setAttack_speed(double attack_speed) {
         this.attack_speed = attack_speed;
     }
